feat(api): Test schema fields

+++ packages/api/lib/graphql/types/index.ts
@@ -19,7 +19,7 @@ const typeDefs = gql`
 
   type Trip {
     geoJson: Geometry!
-    distance: Int!
+    distance: Float!
   }
 
   type Mutation {
